requirements.txt
pandas
requests
sqlalchemy
mysql-connector-python
python-dotenv
